ceres-voice-module/requirements.txt
fastapi
uvicorn
websockets
google-generativeai
requests
openai-whisper
numpy
pvporcupine
torch
pyttsx3
pyaudio
torchaudio
dotenv
